LayoutTests/imported/w3c/web-platform-tests/css/css-values/attr-argument-grammar-expected.txt


FAIL CSS Values and Units Test: attr() argument grammar assert_equals: Value 'attr(var(--x))', where 'data-foo=10' should be valid for the property 'font-weight'. expected "10" but got "400"
FAIL CSS Values and Units Test: attr() argument grammar 1 assert_equals: Value 'attr(var(--x),)', where 'data-foo=10' should be valid for the property 'font-weight'. expected "10" but got "400"
FAIL CSS Values and Units Test: attr() argument grammar 2 assert_equals: Value 'attr(var(--x), 10)', where 'data-foo=invalid' should be valid for the property 'font-weight'. expected "10" but got "400"
PASS CSS Values and Units Test: attr() argument grammar 3
PASS CSS Values and Units Test: attr() argument grammar 4
PASS CSS Values and Units Test: attr() argument grammar 5
PASS CSS Values and Units Test: attr() argument grammar 6
PASS CSS Values and Units Test: attr() argument grammar 7
PASS CSS Values and Units Test: attr() argument grammar 8
FAIL CSS Values and Units Test: attr() argument grammar 9 assert_equals: Value 'attr(var(--y), 3px)', where 'data-foo=var(--y)' should be valid for the property '--x'. expected "3px" but got "data-foo type(<number>), 10"
PASS CSS Values and Units Test: attr() argument grammar 10
PASS CSS Values and Units Test: attr() argument grammar 11
